:bug: Update consent clientside
